Move commonly used test helpers into controllers_utils_test.go file
